Finished Todo App (as far as possible by now)
For some reason the checkboxes are not cooperating with me. The value of the variable assigned to them can be changed but for some reason the checkboxes are sometimes working and on the next try they are not. If you know what could be the problem I woud be more than happy about some feedback!